requirements.txt
fastapi
uvicorn
pydantic

nltk
wordcloud
matplotlib
pandas
numpy
scikit-learn

python-dotenv
pyyaml
mlflow